+++ CHANGELOG.md
@@ -28,10 +28,12 @@
 - [Enhancement] Introduce a `/ux` page to ease with styling improvements and components management.
 - [Enhancement] disallow usage of `<script>` blocks to prevent XSS.
 - [Enhancement] Display full subscription group information in the Routing view, including injectable defaults.
+- [Enhancement] Report Karafka consumer server execution mode.
 - [Refactor] Optimize subscription group data tracking flow.
 - [Refactor] Namespace migrations so migrations related to each topic data are in an independent directory.
 - [Refactor] Use errors for deny flow so request denials can occur from the inspection layer.
 - [Maintenance] Require `karafka` `2.4.7` due to fixes and API changes.
+- [Fix] Disallow quiet and stop commands for swarm workers.
 - [Fix] Disallow quiet and stop commands for embedded workers.
 - [Fix] Fix invalid deserialization metadata display in the per-message Explorer view.
 - [Fix] Fix a case where started page refresh would update content despite limiters being in place.